[*] updated the package name
