+++ library/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
   -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="in.excogitation.lib.sensey">
+    package="com.github.nisrulz.sensey">
 
     <application android:allowBackup="true"
         android:supportsRtl="true">
