+++ sample/src/main/res/layout/activity_main.xml
@@ -24,7 +24,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".MainActivity"
+    tools:context="com.github.nisrulz.senseysample.MainActivity"
     >
